requirements.txt
flask
flask-cors
numpy
pandas
faiss-cpu
google-generativeai